# Optimize FigureLength struct conversion to string, reduce allocations