## Description

Optimizes conversion of `FigureLength` to `string`. Simplifies Can* methods, adds some documentation.
- The benchmark is done with #9887 which allows for better perf navigating through the methods.

### Simple ToString conversion

| Method    | figureLength | Mean [ns]   | Error [ns] | StdDev [ns] | Gen0   | Code Size [B] | Allocated [B] |
|---------- |------------- |------------:|-----------:|------------:|-------:|--------------:|--------------:|
| Original  | 999          | 102.3501 ns |  2.0471 ns |   3.0640 ns | 0.0048 |       4,171 B |          80 B |
| PR__EDIT  | 999          |  76.9590 ns |  1.3950 ns |   1.2367 ns | 0.0029 |         957 B |          48 B |

```csharp
yield return new FigureLength(999.0, FigureUnitType.Column);
```

## Customer Impact

Improved performance, decreased allocations.

## Regression

No.

## Testing

Local build, conversion tests.

## Risk

Low; do note that #9887 should be taken first imho.